Add password recovery mechanism
The password recovery mechanism is initiated when the user requests a
token to be sent to email associated with his account. The server will
send him a link containing the token required to set a new password for
his account.

The database for users is also updated with additional
'recover_password_token' and 'recover_password_token_expiry' fields to
store information about password recovery process.