Integrate multi storage providers to GQL schema
We add new StorageProvider type to the schema and add new attributes to
User and Link types. In the User type, we introduce
connectedStorageProviders to list the storage providers the user
connected to. While in Link type, we add nullable storageProvider to
give information about which storage provider that the link will upload
the files to.