add providerId to createLink and updateLink operation
We add providerId to createLink and updateLink operation to give an
information to the link to which storage the file will be uploaded. In
this patch, we also change the updateLink response type from Message to
Link to give the client the updated link directly. We also improve the
Create method of in-memory LinkRepository in order to provide better
unit tests.

Other minor changes are the refactor the mapping process from Link
(domain) to Link (GraphQL model) by introducing both formatLink and
formatLinks helper functions, and comment of exported methods in
resolvers.